Add feature to avoid logging deprecated geolocation histograms.
On local traces, this logging code takes about 50% of the time
spent generating generating geolocation headers, which is fairly
slow to start with.

We use a Feature to disable this rather than simply removing the code
because we want to measure the impact of reducing InputToNavigationStart
metrics. This latency is not tracked in most page load metrics (e.g.
LargestContentfulPaint).

Also remove CacheDeprecatedSystemLocationSetting fieldtrial config
and bundle it with a new OptimizeGeolocationHeaderGeneration config
(both features will be launched together via Finch).

Bug: 1330739
Change-Id: If349e53a95becddcfed77a4dca3f6767aca6d158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3764701
Commit-Queue: Simon Pelchat <spelchat@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1027258}
NOKEYCHECK=True
GitOrigin-RevId: 9aa28454c2c36c9b3f9c726175645c447ee7733b