[Status Chip] Remove feature logic
Remove code since we are not launching it. However, some part of the
feedback is not deleted since we would want to add another way to show
feedback for save card. If necessary I will make another CL cleaning up
further on that part.

Bug: 932818
Change-Id: I70017ea89a8488be6d3e99257cad2b5580edbfaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710120
Commit-Queue: Siyu An <siyua@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1029900}
NOKEYCHECK=True
GitOrigin-RevId: 6d5fbd4c0f8ce4abab719404a6ce54a5ae06a17b