[Test Automation] Method to retrieve screen bounds of a WebUI element
This CL adds test-only utility methods for retrieving the screen bounds of a WebUI element. These bounds can be used to send drag/drop, click, etc. events during an interactive UI test in order to test the behavior of a WebUI page in a live browser with real input.

Tests are provided, including tests on a live existing WebUI page, in which a button deep in the shadow DOM is located and clicked on, and the UI change resulting from the click is observed.

Note that the test which clicks a WebUI element after locating it is disabled on Wayland because all tests which inject mouse moves seem to fail on Wayland (see other disabled tests in the .filter file).

Bug: 1350614
Change-Id: I2ceee223115c935678ec9abc88683ac3b64ac244
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3813699
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1033132}
NOKEYCHECK=True
GitOrigin-RevId: 794e1ca22a5010f8c415bd09cad8dc98e4566c4b