Clean up Webview CTS variants config
This is a no-op as far as the generated JSONs are concerned.
Move CIPD deps configs on the test suites, where they more logically
belong, given they are the same for all variants.

Also, remove an invalid line which has no effect, but if it had,
it would be breaking (attempts to override test name to a 32-bit
specific one on a variant which is also used for a 64-bit suite)

Bug: None
Change-Id: I3c3f6a507c4c8910e6fdb3fa3194bac1455e4321
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3849012
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Auto-Submit: Mateusz Zbikowski <zbikowski@google.com>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/main@{#1038248}
NOKEYCHECK=True
GitOrigin-RevId: 3fb397440a277329a1cb6d77b510c24bbb0c5f54