[omnibox] [bookmark-paths] `MatchTitledUrlNodeWithQuery()` performance.
tldr; before doing the real check whether an input matches a bookmark,
1st do a quicker approximate check.

`MatchTitledUrlNodeWithQuery()` is a main latency contributor to
bookmark matching for the omnibox. With the bk paths feature, it's by
far the main contributor; with the feature disabled; it's tied with
`GetTypedCountForUrls()`.

The latency comes mainly from checking whether a bk node matches the
input. This requires calling `ExtractQueryWords()` on the title and URL
of every potential bk match.

To reduce the cost, this CL adds a pre-check before doing the real check
using `ExtractQueryWords()`. The pre-check whether every input term is
in either the bk title or URL. Unlike the real check, it doesn't check
the match is at a word break or adheres to the exact match (i.e. quoted
texts) requested by the query. Therefore, the pre-check will have
false-positives, but not false-negatives; if the pre-check passes, we
still do the real check afterwards, but if the pre-check fails, we avoid
the real check.

The pre-check is done only if the bk-path feature is enabled, as it's
guaranteed to pass otherwise.

This is feature guarded by `OmniboxBookmarkApproximateNodeMatch`.

I also considered, lazy-caching the query words for each node's title
and URL, and though that was promising from local testing, it was messy
and might have memory concerns.

Bug: 1129524, 1252537, 1143217
Change-Id: Id8bc2441c03df0eeddd730af7de5d74349df0850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3867482
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1042865}
NOKEYCHECK=True
GitOrigin-RevId: fba377ea90e738b49e31ab91a445371a0174182a