subframe-reuse: enable by default
This change enables the SubframeShutdownDelay experiment by default.
The experiment has shown reasonable improvements on macOS and Windows
[1].

The experiment was not an improvement on Linux, and the impact was
mostly neutral on ChromeOS (which is Linux-based), so the behavior
will not be enabled by default on those platforms.

Enabling this feature by default will allow us to make iterative
improvements if possible, e.g., experimenting with different subframe-process lifetimes.

[1] https://docs.google.com/document/d/1AtnUmHuf8GqUp1zv49ImZDN59_4I0aT_WSEf4wibIt0

Bug: 894253
Change-Id: I34cc6849d207b2783b33b0241f4d54a20fa47bca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3811446
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/main@{#1045166}
NOKEYCHECK=True
GitOrigin-RevId: e8f92ef60b4de197659b7202348f0162736b5cc7