Add AutofillRefinedPhoneNumberTypes to field trial config.
This feature consists of 2 arms (+ Disabled). As only a single arm can
be added to the field trial config, the WithoutPhoneNumberSeparators one
is chosen. This arm doesn't enable the
AutofillConsiderPhoneNumberSeparatorsValidLabels feature, which is
currently only supported on non-iOS platforms*. As the outputs of the
data driven heuristics tests change, enabling this features causes
problems with the iOS bots otherwise.

The updated outputs of the data driven heuristics tests show that the
correct trunk prefix types are assigned.

*The reason is that label inference is implemented separately on iOS and
passing a finch flag there is extremely cumbersome. See
https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/ios/form_util/resources/fill.js;l=1325;drc=3d4f8d32ae06c684c133328a958be17f0b70d1d0.
If the feature is successful, this 1 line regex update can be done on
iOS as well.

Bug: 1311937
Change-Id: I386959dd14c4e74bd1c572796b3aab655190d233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3876376
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046826}
NOKEYCHECK=True
GitOrigin-RevId: 9357223b151ddeb10ba3113a8a2db84c8698ade0