[viz] Enable AggressiveFrameCulling and EagerSurfaceGarbageCollection on bots
Enable both these features together, as they are somewhat linked
(i.e. on Android, eager collection is required to make aggressive frame
culling work well in background), and also because it is the
configuration we are aiming to launch eventually.

Also fix a typo in the feature name before it's too late.

Bug: 1355820
Change-Id: I1c26d45c7291a3295e90d23439189a24ea7ef3dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3893738
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046916}
NOKEYCHECK=True
GitOrigin-RevId: 63ee0191a370c651ed64991af6417623cc914087