infra: Add trybot {mac,win,android}-perfetto-rel
Note: `linux-perfetto-rel` was already there.

For context: We're getting closer to making the switch to the
Perfetto SDK-backed tracing backend in Chrome, and would like
additional coverage on the waterfall for other major platforms in this
configuration (likely until Q1-Q2 2023, when we hope to have made
the switch by).

Bug: 1362224
Change-Id: Iebceaddd7e94c0cf585ab8339acb3e1fe994c846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3870606
Commit-Queue: Mohit Saini <mohitms@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1048317}
NOKEYCHECK=True
GitOrigin-RevId: 96799c335509c46cccffac49cfb04adaa0b288cc