Remove unused consolidated site storage controls code paths
The consolidated site storage controls changes to settings launched
as part of the Privacy Sandbox Settings 3 launch - which has been
default enabled on ToT.

This CL removes code paths which are no longer active within settings,
including the removal of the site data pages.

The changes to the ChromeContentBrowserClient have been removed without
being solidified, as these changes were _not_ launched. The redirect was
nice to have, but not a requirement for launching.

The "Remove 3P cookies" button components in settings are also
removed. This feature was never launched, and only resided on the
removed site data page. That feature has integrations elsewhere in
the codebase that will be removed separately.

LOW_COVERAGE_REASON=This CL moves around tests, but doesn't change
them, so coverage is not negatively impacted, just revealed.

Change-Id: I8ac7b30c44b78eecf22bc8d5991364d8cc57f8d0
Fixed: 1238757
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3921494
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1057636}
NOKEYCHECK=True
GitOrigin-RevId: a420414c0d4cab8163be12cda0b7b0e7a4bc8d1a