Use smoke test file for skia-vulkan-swiftshader
Smoke test file is now the general mechanism to list the test to run for
a flag specific builder. The rebaseline tool will look into the smoke
test file to know which test is run for a flag specific suite, thus knows
when to not copy a baseline for a flag specific builder.

Also add '--skip=always' for highdpi, so that test can be skipped
through test expectations when smoke test file is used.

Bug: 1373574
Change-Id: I26568174c336262d30c31add9fb00d2398353d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3958193
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Cr-Commit-Position: refs/heads/main@{#1059993}
NOKEYCHECK=True
GitOrigin-RevId: e22f9c3eb0414c9bbe7ae23a4f5f39a2ecb8cb8b