Fenced frames: Enable FencedFramesInnerConfig in fieldtrial config
This is the feature flag that we'll use to incrementally enable
HTMLFencedFrameElement#innerConfig attribute support. It will be
enabled on a subset of the current experimentation population, and will
allow fenced frame consumers to generate WebIDL FencedFrameInnerConfig
objects that the `<fencedframe>` element will be able to "navigate" to.

Bug: 1123606

Change-Id: I30d4e0c991bb75fedb087439b250ba2455fee293
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3983945
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1063981}
NOKEYCHECK=True
GitOrigin-RevId: 77f1f2514ad95fb76bfafb112422b3fad7a09827