Attempt to deflake tests by moving tests to ui_interactive_tests.
After much investigation, I have come to the conclusion that AccessCodeCastHandlerBrowserTests are mostly flaky due to loss of focus. This loss of focus causes the dialog to close at random times during the test -- causing much of the flakiness. This is evidence by the tests flaking on CQ bots that run browsertests in parallel.

The solution to this issue is moving the browsertests to interactive_ui_tests. These tests do NOT run in parallel and have fixed flaky issues on swarming CQ bots (so far...).

Bug: b/249794432, b/235896651
Change-Id: I0845f1c8164d4284b54af3d551a29524d0b4a04f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3988764
Commit-Queue: george benz <gbj@google.com>
Reviewed-by: Benjamin Zielinski <bzielinski@google.com>
Cr-Commit-Position: refs/heads/main@{#1065051}
NOKEYCHECK=True
GitOrigin-RevId: eff639a50895bc1e708ed94e1ec6897397c8ac5b