test_env.py: Skip loading vulkan modules for MSAN
This CL fixes the below MSAN error.  This issue is not triggered on the
bots because they run on an older release of Ubuntu (Xenial) which
lacks loadable vulkan modules (eg libvulkan_radeon.so and
libvulkan_intel.so) which are present on newer systems like Ubuntu Focal
and gLinux.  The system modules cause the error because they are not
instrumented with MSAN.

Uninitialized bytes in __interceptor_strcmp at offset 0 inside [0x7fffffffa380, 8)
[Detaching after fork from child process 2199490]
    #0 0x7fffee94014c in wsi_unsupported_instance_extension wsi.c
    #1 0x7fffee8c15f0 in loader_add_instance_extensions loader.c
    #2 0x7fffee8c0c77 in loader_get_icd_loader_instance_extensions third_party/vulkan-deps/vulkan-loader/src/loader/loader.c:1116:15
    #3 0x7fffee8fbe2c in vkCreateInstance third_party/vulkan-deps/vulkan-loader/src/loader/trampoline.c:556:11
    ...

Change-Id: I2cfbd9404a0d31032b1c9b62929215eaf6b4ca26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4007275
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1068202}
NOKEYCHECK=True
GitOrigin-RevId: 509e6712e197c9461ccbbc8e801532e6f811b228