Enable kRegionCaptureExperimentalSubtypes by default
The feature was gradually rolled out and caused no stability issues.
However, that is not surprising, as few developers (if any) would have
had the time to notice this change and take advantage of it.
As such, the feature flag is retained.

Bug: 1332641, 1382005
Change-Id: I5d7d5768b1ad1a95933286c579c3f1081a601ce2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4008319
Auto-Submit: Elad Alon <eladalon@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1069205}
NOKEYCHECK=True
GitOrigin-RevId: 4548a8328174f8d590e5a94f8b197744530e3566