Extend timeout value on blink-rel bots to 40 minutes
Devs(especially Skia team) sometimes make changes that will require
massive updates to baselines. A lot of tests will fail and bots will
timeout due to a lot of time is used for retry. Rebaseline based on such
results is not reliable and is causing pain to developers.

Extend the timeout value significantly to handle such scenarios. The
timeout value on CQ bots will not be changed, so that overall run time
for blink_*_tests will still be bounded by that.

Next step is to add all flag specific steps to linux-blink-rel, and we
should only use blink-rel bots to rebaseline in future.

Bug: 1381873, 1379583
Change-Id: I39eca3a6f0b0ae206a935179443aeda6b57c41cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4009808
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1069442}
NOKEYCHECK=True
GitOrigin-RevId: 21063c77fe5892bfd7bc494aa2097d267775c173