site_isolation: Enable CacheSiteIsolationMemoryThreshold by default
No noticeable results on Stable. The change is good according to local
profiling, probably the effect is just too small, and this overhead is
not a significant bottleneck.

Setting the feature by default, but not removing it yet. Suggesting to
keep the base::Feature for a holdback. We have a bag of small tweaks
running disabled as a holdback in AndroidScrollOptimizations.

More details about the data from the field can be found at
http://crbug.com/1346663#c24.

Bug: 1346663
Change-Id: I0a2114c1f3978c69201c6491a4ef5ba7facd26b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4027911
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1071752}
NOKEYCHECK=True
GitOrigin-RevId: 6ff9c5949ad94479869210feca430b87e6d1fd72