Stop blocking BFCache on App-Banner.
It's in stable at 100%

This converts the chrome browser test to no longer be parameterized on BFCache enabled or not. BFCache is enabled by default and we have a bot which runs with BFCache disabled. This test will verify correctness for both options.

This removes the content browsertest entirely since it was flaky on
Mac and redundant if we have the chrome test.

This leaves the reasons in the enum. They are removed in a follow-on just to keep the numbers of OWNERS reasonable.

Fixed: 1276864
Change-Id: If065d718d8b3996e853dea9d0b82109003b14804
Cq-Include-Trybots: luci.chromium.try:linux-bfcache-rel,android-bfcache-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3912297
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Auto-Submit: Fergal Daly <fergal@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1072492}
NOKEYCHECK=True
GitOrigin-RevId: 6076c47ae864e887191f3b9936300ed9bf3661ce