Remove use of the mixed_target template for Rust Gtest unit tests
mixed_target is a large complex tool with circular dependencies between
Rust and C++ and it is not needed. All we need is a way to generate a
rust_static_library() with a given set of sources and deps, and link it
into the unit test binary. We can do that more simply by just making
the GN rule in testing/test.gni.

This removes the last use of mixed_target.

While doing so, we simplify the use of Rust from test() targets, by
allowing the target to simply add Rust source files to `sources` and
dependencies to `deps`.

This change points out a bug that a rust_target is unable to be
depended on by C++ at the moment if it uses the stdlib, only a mixed
target can, and this is fixed.

Bug: 1381552
Change-Id: I39912415e77beb3c15272922a5f1f6c3425b8804
Cq-Include-Trybots: luci.chromium.try:linux-rust-x64-rel,linux-rust-x64-dbg,android-rust-arm-dbg
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4032367
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073419}
NOKEYCHECK=True
GitOrigin-RevId: 482580a349b7346a1afc1a52ec7b102b300a0427