[omnibox][bookmark-paths] Remove dcheck for removing paths form index.
We index words in bookmark folder names in a `map<word, count>` index in
order to quickly determine if a word is in any folder name. E.g., if we
have 2 bk folders 'red cat' and 'red kitten', the index will be:
`{'red', 2}, {'cat', 1}, {'kitten', 1}}`.

Tracking the counts allows us to correctly remove words from the index
when folders are renamed or deleted. Doing so 1st DCHECKs that the word
existed in the map to begin with. This works in real runtime (i.e. not
tests) since you can't delete or rename a folder that didn't already
exist.

But this doesn't work the `ChunkedPolicyPrefsTest.PolicyToPrefsMapping`
browser test, which deals with managed bookmark folders. It loads the
managed bookmark folder differently than real runtime would, causing the
DCHECK to fire. Changing it ended up being too complicated and kind of
low value since the test doesn't really care about bookmarks or their
indexes.

So this test instead removes the `DCHECK`.

Bug: 1129524
Change-Id: If569f1317788fe74d965b5446fb13fab77d75f89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4076949
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078853}
NOKEYCHECK=True
GitOrigin-RevId: e93dc6ab63c0ed24e339eb88d9182751dcbc0b98