Make freeze violations warnings for presubmit --all
This is a follow-on to crrev.com/c/4115566.

The tools/mb and testing/buildbot freeze violation errors were going to
make the presubmit --all bot red through the entire freeze timeframe so
this change makes them just be a warning when presubmit --all or
presubmit --files is invoked. This allows full testing of the freeze
code, without breaking the bot.

This failure was initially confusing because there was no matching entry
in the blamelist, but that makes sense in hindsight.

The bot is at:
    https://ci.chromium.org/p/chromium/builders/ci/linux-presubmit

Because this change violates itself (how circular) it can only be landed
by putting a tag in the commit message telling the presubmit system to
ignore these failures. This should generally not be allowed/needed, but
this is an exception.

Ignore-Freeze: true
Bug: 1322936
Change-Id: Ide884fcd948a3e5075dece3f841f448e16218f11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4116240
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085539}
NOKEYCHECK=True
GitOrigin-RevId: 703419372930b4d5842b9ec78c3cb6051b839e80