xvfb.py: start Weston with custom config disabling the panel
Weston only allows limited configuration via CLI flags. To ensure that
the panel does not interfere with tests by blocking mouse input, we need
to disable it in the config file.

Please note that the panel should currently not be causing any problems,
but it might do so in the future when we switch to using the new
ui-controls protocol instead of weston test, as that allows manual
positioning of windows. Instead of re-writing tests to make sure windows
aren't positioned under the panel, we just disable the panel completely,
as we don't need it for the tests.

Bug: 1353089
Change-Id: I4d82bea65b2e4fb9792fb500544387cbca64e419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4116782
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1090274}
NOKEYCHECK=True
GitOrigin-RevId: d0925c3c94ff4d42db90c6e924866b13928e9154