WebGPU: Re-remove testing with validation layers on Intel bots
Testing was originally disabled in crrev.com/c/3584408 due to lack of
capacity.
It was re-enabled in crrev.com/c/4134064 in the hope that new Optiplex
7080 pool will have enough capacity.
Unfortunately, this turned out to be not the case, so these tests need
to be disabled again.

Bug: 1406194, 1349828
Change-Id: Ib2879304a6c17cbb590c6958d03136df0f93d96c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4150311
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Auto-Submit: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1090822}
NOKEYCHECK=True
GitOrigin-RevId: 9f43dc5e64c6f8d9345d64a30609bc9298eee6ba