[Fuchsia] Reduce the test parallelism of content_browsertests on fuchsia-x64-cast-receiver-rel to 1
A fuchsia kernel panic drops the test log if the test case is hanging.
But currently we are running two tests in parallel, it's hard for us to
determine which one causes the issue before the kernel panic is fixed.
So this is a temporary workaround to limit the parallelism of
content_browsertests to be 1 so that we can tell which test case is
flaky.

Search
`Using 1 parallel jobs.`
in https://chromium-swarm.appspot.com/task?id=5fc19678bec49610
to demonstrate the impact.

Bug: 1406693
Change-Id: Ibeff5171889793c74c8430b96171b7b69ccff199
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4160529
Auto-Submit: Zijie He <zijiehe@google.com>
Reviewed-by: Chong Gu <chonggu@google.com>
Commit-Queue: Chong Gu <chonggu@google.com>
Cr-Commit-Position: refs/heads/main@{#1092392}
NOKEYCHECK=True
GitOrigin-RevId: 0f361e65a1a8f2e65582bb8dde1c7ff20eb9928e