Finch: Run commands for disabling or enabling airplane mode on devices in the Android test suite
This CL updates methods in the finch smoke tests suite that disable
or enable internet on the test device. It adds commands to disable
or enable airplane mode to these methods. This is one of the changes
necessary for the Android finch test suite to support API 24.

Bug: b/266222337
Change-Id: I3c7fd16f8cca9a5a283c039af9e2293ef70a2e84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4197658
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/main@{#1098681}
NOKEYCHECK=True
GitOrigin-RevId: 9799780bb8e96d77eeaf227894dd8565bbe13687