Extend timeout for zlib_unittests to 400s
The default timeout is 45s and this is hitting for ZipTest.BigFile.

This CL extends it to 400 seconds. This is based on the calculation
that the test typically only reaches 15% in 45 seconds, so should
reach 100% in 300 seconds, with an extra 100 seconds to spare.

BUG=1299736, 1413911

Change-Id: I4705d1853a796b7748e307cb250460f2837d427d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4229952
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103105}
NOKEYCHECK=True
GitOrigin-RevId: d85f20699389e4c457b0a3256525d70e948c1396