Remove PercentDroppedFrames v1
The v1 of this metric is no longer needed. ChromeOS has begun using v3.
Experiments relying on it have wrapped up, or moved on to v3.

With v3 being a guardrail metric this is no longer needed.

Also remove the UKMs for this. We currently do not plan to do further
analysis here. Nor do we have plans to add v3.

Bug: 1405963
Change-Id: Idffacdfb15e3a06237dd88e86c3524e8ac31f00d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210171
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Mingjing Zhang <mjzhang@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103450}
NOKEYCHECK=True
GitOrigin-RevId: 41b7b34ab617ae83a7a31eec431cdbf9fbbdadd4