[MerchantTrust] Post-launch cleanup
The feature is fully launched in US without showing the message. We
add a "launched" flag "CommerceMerchantViewerRegionLaunched" which is
used in combination with the country and locale checks to determine
whether the feature is launched in a particular country.

We also add the JNI bridge so we can use the check in Java.

Another related feature flag "PageInfoStoreInfo" is enabled by
default without country and locale checks since it neither shows
anything to users nor initiates any server calls when the
"CommerceMerchantViewer" flag is off.

Low-Coverage-Reason: mock_shopping_service is for test only.

Bug: 1400834
Change-Id: Id8db0ef257414afdfa0f22d163ed252e19f23549
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4144465
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Ayman Almadhoun <ayman@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Zhiyuan Cai <zhiyuancai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1104817}
NOKEYCHECK=True
GitOrigin-RevId: 50efe8a5cc64f83ee18e4ef25a9d6c52e800fa5c